# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  memory.AddRequest:
    properties:
      agent_id:
        type: string
      filters:
        additionalProperties: true
        type: object
      infer:
        type: boolean
      message:
        type: string
      messages:
        items:
          $ref: '#/definitions/memory.Message'
        type: array
      metadata:
        additionalProperties: true
        type: object
      run_id:
        type: string
      user_id:
        type: string
    type: object
  memory.DeleteAllRequest:
    properties:
      agent_id:
        type: string
      run_id:
        type: string
      user_id:
        type: string
    type: object
  memory.DeleteResponse:
    properties:
      message:
        type: string
    type: object
  memory.MemoryItem:
    properties:
      agentId:
        type: string
      createdAt:
        type: string
      hash:
        type: string
      id:
        type: string
      memory:
        type: string
      metadata:
        additionalProperties: true
        type: object
      runId:
        type: string
      score:
        type: number
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  memory.Message:
    properties:
      content:
        type: string
      role:
        type: string
    type: object
  memory.SearchRequest:
    properties:
      agent_id:
        type: string
      filters:
        additionalProperties: true
        type: object
      limit:
        type: integer
      query:
        type: string
      run_id:
        type: string
      user_id:
        type: string
    type: object
  memory.SearchResponse:
    properties:
      relations:
        items: {}
        type: array
      results:
        items:
          $ref: '#/definitions/memory.MemoryItem'
        type: array
    type: object
  memory.UpdateRequest:
    properties:
      memory:
        type: string
      memory_id:
        type: string
    type: object
info:
  contact: {}
  description: User-scoped filesystem API for containerd-backed data.
  title: Memoh Go API
  version: "1.0"
paths:
...
  /memory/add:
    post:
      description: Add memory for a user via memory
      parameters:
      - description: Add request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/memory.AddRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.SearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Add memory
      tags:
      - memory
  /memory/memories:
    delete:
      description: Delete all memories for a user via memory
      parameters:
      - description: Delete all request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/memory.DeleteAllRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.DeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete memories
      tags:
      - memory
    get:
      description: List memories for a user via memory
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Agent ID
        in: query
        name: agent_id
        type: string
      - description: Run ID
        in: query
        name: run_id
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.SearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List memories
      tags:
      - memory
  /memory/memories/{memoryId}:
    delete:
      description: Delete a memory by ID via memory
      parameters:
      - description: Memory ID
        in: path
        name: memoryId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.DeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete memory
      tags:
      - memory
    get:
      description: Get a memory by ID via memory
      parameters:
      - description: Memory ID
        in: path
        name: memoryId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.MemoryItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get memory
      tags:
      - memory
  /memory/search:
    post:
      description: Search memories for a user via memory
      parameters:
      - description: Search request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/memory.SearchRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.SearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Search memories
      tags:
      - memory
  /memory/update:
    post:
      description: Update a memory by ID via memory
      parameters:
      - description: Update request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/memory.UpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.MemoryItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update memory
      tags:
      - memory
schemes:
- http
swagger: "2.0"